fixed last fix

+++ output/main.js
@@ -1368,7 +1368,8 @@ var figObjectArrays = new Array(); // [ {nodeId, [objects]} ]
 var figStyleArrays = new Array(); // [ {nodeId, [styles]}  ]
 function figDeleteObjectsFromNodeIds(nodeIds) {
     for (let i = figPoints.length - 1; i >= 0; i--)
-        if (nodeIds.includes(figPoints[i].getPluginData('nodeId')))
+        if (!figPoints[i].removed
+            && nodeIds.includes(figPoints[i].getPluginData('nodeId')))
             figPoints.splice(i, 1);
     for (let i = figEmptyObjects.length - 1; i >= 0; i--)
         if (figEmptyObjects[i].removed
@@ -1382,6 +1383,8 @@ function figDeleteObjectsFromNodeIds(nodeIds) {
 }
 function figDeleteAllObjects(forceDelete = false) {
     for (const obj of figma.currentPage.children) {
+        if (obj.removed)
+            continue;
         if (obj.getPluginData('objectId') != ''
             && (obj.getPluginData('retain') == '0'
                 || forceDelete)
@@ -2333,7 +2336,8 @@ function figUpdateObjects(figParent, genObjects, nodeIds = [], firstChunk = fals
                 || figObj == null
                 || figObj.removed) // no existing object, create new one
                 yield figCreateObject(genObj, addObject);
-            else if (figObj.getPluginData('type') == genObj[FO_TYPE].toString()) // update existing object
+            else if (!figObj.removed
+                && figObj.getPluginData('type') == genObj[FO_TYPE].toString()) // update existing object
                 yield figUpdateObject(figObj, genObj);
             else // delete existing object, create new one
              {
@@ -2350,7 +2354,7 @@ function figUpdateObjects(figParent, genObjects, nodeIds = [], firstChunk = fals
             && !figParent.removed) {
             for (const figObj of figParent.children)
                 if (figObj.removed
-                    && !genObjects.find(o => o[FO_OBJECT_ID] == figObj.getPluginData('objectId')))
+                    || !genObjects.find(o => o[FO_OBJECT_ID] == figObj.getPluginData('objectId')))
                     figObj.remove();
         }
         // put points on top
